+++ src/lib/config.ts
@@ -61,9 +61,9 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 
 export const SDK_METADATA = {
   language: "typescript",
-  openapiDocVersion: "4.14.0-837595d5",
-  sdkVersion: "0.0.54",
+  openapiDocVersion: "4.15.0-alpha.1758566860656-1c3a24b3",
+  sdkVersion: "0.0.55-alpha.1",
   genVersion: "2.707.0",
   userAgent:
-    "speakeasy-sdk/typescript 0.0.54 2.707.0 4.14.0-837595d5 cribl-control-plane",
+    "speakeasy-sdk/typescript 0.0.55-alpha.1 2.707.0 4.15.0-alpha.1758566860656-1c3a24b3 cribl-control-plane",
 } as const;