+++ src/models/outputdatadog.ts
@@ -4,7 +4,12 @@
 
 import * as z from "zod";
 import { safeParse } from "../lib/schemas.js";
-import { ClosedEnum } from "../types/enums.js";
+import {
+  catchUnrecognizedEnum,
+  ClosedEnum,
+  OpenEnum,
+  Unrecognized,
+} from "../types/enums.js";
 import { Result as SafeParseResult } from "../types/fp.js";
 import { SDKValidationError } from "./errors/sdkvalidationerror.js";
 
@@ -23,7 +28,7 @@ export const SendLogsAs = {
 /**
  * The content type to use when sending logs
  */
-export type SendLogsAs = ClosedEnum<typeof SendLogsAs>;
+export type SendLogsAs = OpenEnum<typeof SendLogsAs>;
 
 /**
  * Default value for message severity. When you send logs as JSON objects, the event's '__severity' field (if set) will override this value.
@@ -41,7 +46,7 @@ export const OutputDatadogSeverity = {
 /**
  * Default value for message severity. When you send logs as JSON objects, the event's '__severity' field (if set) will override this value.
  */
-export type OutputDatadogSeverity = ClosedEnum<typeof OutputDatadogSeverity>;
+export type OutputDatadogSeverity = OpenEnum<typeof OutputDatadogSeverity>;
 
 /**
  * Datadog site to which events should be sent
@@ -58,7 +63,7 @@ export const DatadogSite = {
 /**
  * Datadog site to which events should be sent
  */
-export type DatadogSite = ClosedEnum<typeof DatadogSite>;
+export type DatadogSite = OpenEnum<typeof DatadogSite>;
 
 export type OutputDatadogExtraHttpHeader = {
   name?: string | undefined;
@@ -76,7 +81,7 @@ export const OutputDatadogFailedRequestLoggingMode = {
 /**
  * Data to log when a request fails. All headers are redacted by default, unless listed as safe headers below.
  */
-export type OutputDatadogFailedRequestLoggingMode = ClosedEnum<
+export type OutputDatadogFailedRequestLoggingMode = OpenEnum<
   typeof OutputDatadogFailedRequestLoggingMode
 >;
 
@@ -126,7 +131,7 @@ export const OutputDatadogBackpressureBehavior = {
 /**
  * How to handle events when all receivers are exerting backpressure
  */
-export type OutputDatadogBackpressureBehavior = ClosedEnum<
+export type OutputDatadogBackpressureBehavior = OpenEnum<
   typeof OutputDatadogBackpressureBehavior
 >;
 
@@ -140,7 +145,7 @@ export const OutputDatadogAuthenticationMethod = {
 /**
  * Enter API key directly, or select a stored secret
  */
-export type OutputDatadogAuthenticationMethod = ClosedEnum<
+export type OutputDatadogAuthenticationMethod = OpenEnum<
   typeof OutputDatadogAuthenticationMethod
 >;
 
@@ -154,7 +159,7 @@ export const OutputDatadogCompression = {
 /**
  * Codec to use to compress the persisted data
  */
-export type OutputDatadogCompression = ClosedEnum<
+export type OutputDatadogCompression = OpenEnum<
   typeof OutputDatadogCompression
 >;
 
@@ -168,7 +173,7 @@ export const OutputDatadogQueueFullBehavior = {
 /**
  * How to handle events when the queue is exerting backpressure (full capacity or low disk). 'Block' is the same behavior as non-PQ blocking. 'Drop new data' throws away incoming data, while leaving the contents of the PQ unchanged.
  */
-export type OutputDatadogQueueFullBehavior = ClosedEnum<
+export type OutputDatadogQueueFullBehavior = OpenEnum<
   typeof OutputDatadogQueueFullBehavior
 >;
 
@@ -183,7 +188,7 @@ export const OutputDatadogMode = {
 /**
  * In Error mode, PQ writes events to the filesystem if the Destination is unavailable. In Backpressure mode, PQ writes events to the filesystem when it detects backpressure from the Destination. In Always On mode, PQ always writes events to the filesystem.
  */
-export type OutputDatadogMode = ClosedEnum<typeof OutputDatadogMode>;
+export type OutputDatadogMode = OpenEnum<typeof OutputDatadogMode>;
 
 export type OutputDatadogPqControls = {};
 
@@ -381,12 +386,25 @@ export namespace OutputDatadogType$ {
 }
 
 /** @internal */
-export const SendLogsAs$inboundSchema: z.ZodNativeEnum<typeof SendLogsAs> = z
-  .nativeEnum(SendLogsAs);
+export const SendLogsAs$inboundSchema: z.ZodType<
+  SendLogsAs,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(SendLogsAs),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const SendLogsAs$outboundSchema: z.ZodNativeEnum<typeof SendLogsAs> =
-  SendLogsAs$inboundSchema;
+export const SendLogsAs$outboundSchema: z.ZodType<
+  SendLogsAs,
+  z.ZodTypeDef,
+  SendLogsAs
+> = z.union([
+  z.nativeEnum(SendLogsAs),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal
@@ -400,14 +418,25 @@ export namespace SendLogsAs$ {
 }
 
 /** @internal */
-export const OutputDatadogSeverity$inboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogSeverity
-> = z.nativeEnum(OutputDatadogSeverity);
+export const OutputDatadogSeverity$inboundSchema: z.ZodType<
+  OutputDatadogSeverity,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(OutputDatadogSeverity),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const OutputDatadogSeverity$outboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogSeverity
-> = OutputDatadogSeverity$inboundSchema;
+export const OutputDatadogSeverity$outboundSchema: z.ZodType<
+  OutputDatadogSeverity,
+  z.ZodTypeDef,
+  OutputDatadogSeverity
+> = z.union([
+  z.nativeEnum(OutputDatadogSeverity),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal
@@ -421,12 +450,25 @@ export namespace OutputDatadogSeverity$ {
 }
 
 /** @internal */
-export const DatadogSite$inboundSchema: z.ZodNativeEnum<typeof DatadogSite> = z
-  .nativeEnum(DatadogSite);
+export const DatadogSite$inboundSchema: z.ZodType<
+  DatadogSite,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(DatadogSite),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const DatadogSite$outboundSchema: z.ZodNativeEnum<typeof DatadogSite> =
-  DatadogSite$inboundSchema;
+export const DatadogSite$outboundSchema: z.ZodType<
+  DatadogSite,
+  z.ZodTypeDef,
+  DatadogSite
+> = z.union([
+  z.nativeEnum(DatadogSite),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal
@@ -499,15 +541,25 @@ export function outputDatadogExtraHttpHeaderFromJSON(
 }
 
 /** @internal */
-export const OutputDatadogFailedRequestLoggingMode$inboundSchema:
-  z.ZodNativeEnum<typeof OutputDatadogFailedRequestLoggingMode> = z.nativeEnum(
-    OutputDatadogFailedRequestLoggingMode,
-  );
+export const OutputDatadogFailedRequestLoggingMode$inboundSchema: z.ZodType<
+  OutputDatadogFailedRequestLoggingMode,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(OutputDatadogFailedRequestLoggingMode),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const OutputDatadogFailedRequestLoggingMode$outboundSchema:
-  z.ZodNativeEnum<typeof OutputDatadogFailedRequestLoggingMode> =
-    OutputDatadogFailedRequestLoggingMode$inboundSchema;
+export const OutputDatadogFailedRequestLoggingMode$outboundSchema: z.ZodType<
+  OutputDatadogFailedRequestLoggingMode,
+  z.ZodTypeDef,
+  OutputDatadogFailedRequestLoggingMode
+> = z.union([
+  z.nativeEnum(OutputDatadogFailedRequestLoggingMode),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal
@@ -655,14 +707,25 @@ export function outputDatadogTimeoutRetrySettingsFromJSON(
 }
 
 /** @internal */
-export const OutputDatadogBackpressureBehavior$inboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogBackpressureBehavior
-> = z.nativeEnum(OutputDatadogBackpressureBehavior);
+export const OutputDatadogBackpressureBehavior$inboundSchema: z.ZodType<
+  OutputDatadogBackpressureBehavior,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(OutputDatadogBackpressureBehavior),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const OutputDatadogBackpressureBehavior$outboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogBackpressureBehavior
-> = OutputDatadogBackpressureBehavior$inboundSchema;
+export const OutputDatadogBackpressureBehavior$outboundSchema: z.ZodType<
+  OutputDatadogBackpressureBehavior,
+  z.ZodTypeDef,
+  OutputDatadogBackpressureBehavior
+> = z.union([
+  z.nativeEnum(OutputDatadogBackpressureBehavior),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal
@@ -677,14 +740,25 @@ export namespace OutputDatadogBackpressureBehavior$ {
 }
 
 /** @internal */
-export const OutputDatadogAuthenticationMethod$inboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogAuthenticationMethod
-> = z.nativeEnum(OutputDatadogAuthenticationMethod);
+export const OutputDatadogAuthenticationMethod$inboundSchema: z.ZodType<
+  OutputDatadogAuthenticationMethod,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(OutputDatadogAuthenticationMethod),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const OutputDatadogAuthenticationMethod$outboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogAuthenticationMethod
-> = OutputDatadogAuthenticationMethod$inboundSchema;
+export const OutputDatadogAuthenticationMethod$outboundSchema: z.ZodType<
+  OutputDatadogAuthenticationMethod,
+  z.ZodTypeDef,
+  OutputDatadogAuthenticationMethod
+> = z.union([
+  z.nativeEnum(OutputDatadogAuthenticationMethod),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal
@@ -699,14 +773,25 @@ export namespace OutputDatadogAuthenticationMethod$ {
 }
 
 /** @internal */
-export const OutputDatadogCompression$inboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogCompression
-> = z.nativeEnum(OutputDatadogCompression);
+export const OutputDatadogCompression$inboundSchema: z.ZodType<
+  OutputDatadogCompression,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(OutputDatadogCompression),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const OutputDatadogCompression$outboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogCompression
-> = OutputDatadogCompression$inboundSchema;
+export const OutputDatadogCompression$outboundSchema: z.ZodType<
+  OutputDatadogCompression,
+  z.ZodTypeDef,
+  OutputDatadogCompression
+> = z.union([
+  z.nativeEnum(OutputDatadogCompression),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal
@@ -720,14 +805,25 @@ export namespace OutputDatadogCompression$ {
 }
 
 /** @internal */
-export const OutputDatadogQueueFullBehavior$inboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogQueueFullBehavior
-> = z.nativeEnum(OutputDatadogQueueFullBehavior);
+export const OutputDatadogQueueFullBehavior$inboundSchema: z.ZodType<
+  OutputDatadogQueueFullBehavior,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(OutputDatadogQueueFullBehavior),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const OutputDatadogQueueFullBehavior$outboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogQueueFullBehavior
-> = OutputDatadogQueueFullBehavior$inboundSchema;
+export const OutputDatadogQueueFullBehavior$outboundSchema: z.ZodType<
+  OutputDatadogQueueFullBehavior,
+  z.ZodTypeDef,
+  OutputDatadogQueueFullBehavior
+> = z.union([
+  z.nativeEnum(OutputDatadogQueueFullBehavior),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal
@@ -741,14 +837,25 @@ export namespace OutputDatadogQueueFullBehavior$ {
 }
 
 /** @internal */
-export const OutputDatadogMode$inboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogMode
-> = z.nativeEnum(OutputDatadogMode);
+export const OutputDatadogMode$inboundSchema: z.ZodType<
+  OutputDatadogMode,
+  z.ZodTypeDef,
+  unknown
+> = z
+  .union([
+    z.nativeEnum(OutputDatadogMode),
+    z.string().transform(catchUnrecognizedEnum),
+  ]);
 
 /** @internal */
-export const OutputDatadogMode$outboundSchema: z.ZodNativeEnum<
-  typeof OutputDatadogMode
-> = OutputDatadogMode$inboundSchema;
+export const OutputDatadogMode$outboundSchema: z.ZodType<
+  OutputDatadogMode,
+  z.ZodTypeDef,
+  OutputDatadogMode
+> = z.union([
+  z.nativeEnum(OutputDatadogMode),
+  z.string().and(z.custom<Unrecognized<string>>()),
+]);
 
 /**
  * @internal