+++ RELEASES.md
@@ -499,3 +499,13 @@ Based on:
 - [typescript v0.0.54] .
 ### Releases
 - [NPM v0.0.54] https://www.npmjs.com/package/cribl-control-plane/v/0.0.54 - .
+
+## 2025-09-22 19:02:16
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.621.0 (2.707.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.0.55-alpha.1] .
+### Releases
+- [NPM v0.0.55-alpha.1] https://www.npmjs.com/package/cribl-control-plane/v/0.0.55-alpha.1 - .